Explicitly mark dead code as allowed, where needed
This will silence warnings from the compiler. This in turn will make
any new warnings easier to spot, which is good, in case they're
actually important. Dead code warnings are not particularly important
as such, I think, but if there's many of them, they let other
warnings hide in the asteroid field.